+++ adls/lib/LoggingPolicyFactory.ts
@@ -1,8 +1,4 @@
-import {
-  RequestPolicy,
-  RequestPolicyFactory,
-  RequestPolicyOptions
-} from "ms-rest-js";
+import { RequestPolicy, RequestPolicyFactory, RequestPolicyOptions } from "@azure/ms-rest-js";
 
 import { LoggingPolicy } from "./policies/LoggingPolicy";