+++ adls/lib/Pipeline.ts
@@ -9,8 +9,8 @@ import {
   RequestPolicyFactory,
   RequestPolicyOptions,
   ServiceClientOptions,
-  WebResource
-} from "ms-rest-js";
+  WebResource,
+} from "@azure/ms-rest-js";
 
 // Export following interfaces and types for customers who want to implement their
 // own RequestPolicy or HTTPClient