+++ adls/lib/credentials/SharedKeyCredential.ts
@@ -1,5 +1,5 @@
+import { RequestPolicy, RequestPolicyOptions } from "@azure/ms-rest-js";
 import * as Crypto from "crypto";
-import { RequestPolicy, RequestPolicyOptions } from "ms-rest-js";
 
 import { SharedKeyCredentialPolicy } from "../policies/SharedKeyCredentialPolicy";
 import { Credential } from "./Credential";