+++ adls/lib/policies/BrowserPolicy.ts
@@ -4,8 +4,8 @@ import {
   isNode,
   RequestPolicy,
   RequestPolicyOptions,
-  WebResource
-} from "ms-rest-js";
+  WebResource,
+} from "@azure/ms-rest-js";
 
 import { HeaderConstants, URLConstants } from "../utils/constants";
 import { setURLParameter } from "../utils/utils.common";