+++ adls/lib/policies/UniqueRequestIDPolicy.ts
@@ -4,8 +4,8 @@ import {
   HttpOperationResponse,
   RequestPolicy,
   RequestPolicyOptions,
-  WebResource
-} from "ms-rest-js";
+  WebResource,
+} from "@azure/ms-rest-js";
 
 import { HeaderConstants } from "../utils/constants";