+++ adls/lib/utils/RetriableReadableStream.ts
@@ -1,5 +1,6 @@
-import { RestError, TransferProgressEvent } from "ms-rest-js";
+import { RestError, TransferProgressEvent } from "@azure/ms-rest-js";
 import { Readable } from "stream";
+
 import { Aborter } from "../Aborter";
 
 export type ReadableStreamGetter = (